+++ .warp/bin/mysql_help.sh
@@ -62,10 +62,12 @@ function mysql_dump_help()
     warp_message ""
     warp_message_info "Help:"
     warp_message " Create a backup of a database and save it local machine"
+    warp_message " to remove all the security definers of a mysql dump, add this to the command: sed -e 's/DEFINER[ ]*=[ ]*[^*]*\*/\*/'"
     warp_message ""
 
     warp_message_info "Example:"
     warp_message " warp mysql dump warp_db | gzip > /path/to/save/backup/warp_db.sql.gz"
+    warp_message " warp mysql dump warp_db | sed -e 's/DEFINER[ ]*=[ ]*[^*]*\*/\*/' | gzip > /path/to/save/backup/warp_db.sql.gz"
     warp_message " warp mysql dump warp_db | gzip | pv > /path/to/save/backup/warp_db.sql.gz"
     warp_message " warp mysql dump warp_db > /path/to/save/backup/warp_db.sql"
     warp_message ""