+++ .warp/bin/rsync.sh
@@ -2,12 +2,12 @@
 
     # IMPORT HELP
 
-    . "$PROJECTPATH/.warp/bin/sync_help.sh"
+    . "$PROJECTPATH/.warp/bin/rsync_help.sh"
 
-function push_to_container() {
+function rsync_push_to_container() {
 
   if [ "$1" = "-h" ] || [ "$1" = "--help" ] ; then      
-      push_help_usage
+      rsync_push_help_usage
       exit 0;
   fi
 
@@ -18,45 +18,33 @@ function push_to_container() {
     exit 1;
   fi
 
-  case "$(uname -s)" in
-    Linux)
-      warp_message_error "these commands are available only on mac"
-      exit 1;
-    ;;
-  esac
+  # Check rsync is installed
+  hash rsync 2>/dev/null || warp_rsync_is_not_installed
 
   [ -z "$1" ] && warp_message_error "Please specify a directory or file to copy to container (ex. vendor, --all)" && exit
-  CONTAINER_PHP_NAME=$(docker-compose -f $DOCKERCOMPOSEFILE ps -q php)
+  CONTAINER_APPDATA_PORT=$(docker inspect --format='{{(index (index .NetworkSettings.Ports "873/tcp") 0).HostPort}}' $(warp docker ps -q appdata))
 
   if [ "$1" == "--all" ]; then
-    docker cp ./ $CONTAINER_PHP_NAME:/var/www/html
+    rsync -aogvEh * --executability --chown=$(id -u):33  rsync://localhost:$CONTAINER_APPDATA_PORT/warp
     warp_message "Completed copying all files from host to container"
-    warp fix --owner
   else
     
     for i in "$@"
     do
       if [ -f $i ] || [ -d $i ] ; then
-        docker cp ./$i $CONTAINER_PHP_NAME:/var/www/html
+        rsync -aogvEh $i --executability --chown=$(id -u):33  rsync://localhost:$CONTAINER_APPDATA_PORT/warp
         warp_message "Completed copying $i from host to container"  
       else
         warp_message_error "do not copy $i from host to container"  
       fi
     done;    
-
-    # fix permissions
-    if [ $# -eq 1 ] ; then
-      warp fix --owner $1
-    else
-      [ $# -ge 2 ] && warp fix --owner
-    fi;     
   fi  
 }
 
-function pull_from_container() {
+function rsync_pull_from_container() {
 
   if [ "$1" = "-h" ] || [ "$1" = "--help" ] ; then      
-      pull_help_usage
+      rsync_pull_help_usage
       exit 0;
   fi
 
@@ -67,57 +55,39 @@ function pull_from_container() {
     exit 1;
   fi
 
-  case "$(uname -s)" in
-    Linux)
-      warp_message_error "these commands are available only on mac"
-      exit 1;
-    ;;
-  esac
+  # Check rsync is installed
+  hash rsync 2>/dev/null || warp_rsync_is_not_installed
 
   [ -z "$1" ] && warp_message_error "Please specify a directory or file to copy from container (ex. vendor, --all)" && exit
-  CONTAINER_PHP_NAME=$(docker-compose -f $DOCKERCOMPOSEFILE ps -q php)
+  CONTAINER_APPDATA_PORT=$(docker inspect --format='{{(index (index .NetworkSettings.Ports "873/tcp") 0).HostPort}}' $(warp docker ps -q appdata))
 
   if [ "$1" == "--all" ]; then
-    docker cp $CONTAINER_PHP_NAME:/var/www/html/./ ./
+    rsync -aogvEh rsync://localhost:$CONTAINER_APPDATA_PORT/warp .
     warp_message "Completed copying all files from container to host"
   else
     for i in "$@"
-    do
-        docker cp $CONTAINER_PHP_NAME:/var/www/html/$i ./
+    do  
+        rsync -aogvEh rsync://localhost:$CONTAINER_APPDATA_PORT/warp/$i .
         warp_message "Completed copying $i from container to host"
     done;    
   fi  
 }
 
-function warp_clean_volume()
-{
-  if [ "$1" = "-h" ] || [ "$1" = "--help" ] ; then      
-      clean_help_usage
-      exit 0;
-  fi
-
-  docker-sync clean
-}
-
-function sync_main()
+function rsync_main()
 {
     case "$1" in
         push)
 		      shift 1
-          push_to_container $*  
+          rsync_push_to_container $*  
         ;;
 
         pull)
 		      shift 1
-          pull_from_container $*  
-        ;;
-
-        clean)
-          warp_clean_volume $*
+          rsync_pull_from_container $*  
         ;;
 
         *)
-		      sync_help_usage
+		      rsync_help_usage
         ;;
     esac
 }